requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==4.2.20
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-api-key==2.3.0
idna==3.10
Markdown==3.7
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
